Сделал кучу фиксов и правок по всему коду.

+++ modules/del_mess.py
@@ -1,9 +1,22 @@
 import time
 
+message = None
+num_old = None
 
-def del_mess(message, bot, num):
-	# print('dell_mes')
-	time.sleep(5)
-	for i in range(num):
-		bot.delete_message(message.chat.id, message.message_id - i)
-	return
+def del_mess(message_old, bot, num_old):
+	global message
+	global num
+
+	# time.sleep(30)
+	if message != None:
+		try:
+			for i in range(num):
+				if (message.message_id - i) != None:
+					bot.delete_message(message.chat.id, message.message_id - i)
+
+		except TypeError as e:
+			print('DELETE_ERROR')
+			return
+
+	num = num_old
+	message = message_old
